Move tx hashes gathering hooks to component
Moved hooks to `KnowledgebaseLinks` component. Removed redundant store
slice entries:  `mintingRequestedTxHash` & `mintingFinalizedTxHash`
